fixed logic in random generator

+++ extra/random.h
@@ -39,7 +39,7 @@ class SplitMix64 {
     }
 
     double next_float() {
-        return next_int() / (1 << 64);             /* divide by 2^64 to return a value between 0 and 1 */
+        return ((double)next_int()) / ((double)(1 << 64));             /* divide by 2^64 to return a value between 0 and 1 */
     }
 };
